refactor: Improve naming of components to be legit.
