+++ app/_components/proposalpages/ViewProposal.jsx
